# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This service manages all backend interactions for the feed the realm clients.\n\n⚠️ NOTE: All 2xx responses are wrapped in a 'Data envelope'.\nExample: { \"data\": { \"id\": \"123\", \"name\": \"Alice\" } }",
        "title": "FeedTheRealm Core Service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/assets/cosmetics/worlds/{world_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of cosmetics that belong to a specific world ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assets-service"
                ],
                "summary": "Get cosmetics by world",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "world_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 24,
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.CosmeticsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/assets/items/world/{world_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves an items list specific to a world ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assets-service"
                ],
                "summary": "Get items by world",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "world_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ItemListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload item IDs alongside sprite files mapping to a world.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assets-service"
                ],
                "summary": "Upload items and sprites",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "world_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Array of exact item IDs",
                        "name": "ids",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Muti-part chunk array of files",
                        "name": "sprites",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.ItemListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/assets/materials/world/{world_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload material IDs alongside material files mapping to a world.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assets-service"
                ],
                "summary": "Upload materials",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "world_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Array of exact material IDs",
                        "name": "ids",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Muti-part chunk array of files",
                        "name": "materials",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.MaterialResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/assets/models/world/{world_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit a single custom model bounded to a specific world.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assets-service"
                ],
                "summary": "Upload a 3D model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "world_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Model UUID",
                        "name": "model_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Model GLB file",
                        "name": "model_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.ModelResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/world": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates paginated meta-list of standard player-made worlds.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "List worlds",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by owner user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max hits per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search filters",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.WorldsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new world instance in the registry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Publish new world",
                "parameters": [
                    {
                        "description": "World creation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.WorldRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.WorldResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/world/reset-database": {
            "delete": {
                "description": "Purge utility used carefully during development to wipe all tables.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Development DB reset",
                "responses": {
                    "200": {
                        "description": "Success reset",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CosmeticsListResponse": {
            "type": "object",
            "properties": {
                "cosmetics_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CosmeticResponse"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dtos.ItemListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ItemResponse"
                    }
                }
            }
        },
...
        "dtos.MaterialResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "material_type": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "world_id": {
                    "type": "string"
                }
            }
        },
        "dtos.ModelResponse": {
            "type": "object",
            "properties": {
                "model_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "dtos.WorldRequest": {
            "type": "object",
            "properties": {
                "createable_data": {},
                "data": {},
                "description": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                }
            }
        },
...
        "dtos.WorldsListResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "worlds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.WorldMetadata"
                    }
                }
            }
        },
...
    }
}